netwatch/requirements.txt
# NetWatch — Python package dependencies
# Install with:  pip install -r requirements.txt
#
# NOTE: nmap binary must also be installed separately from https://nmap.org/download.html
#       and added to your system PATH.

# ── Core GUI ────────────────────────────────────────────
PyQt5>=5.15.9

# ── Network scanning ────────────────────────────────────
python-nmap>=0.7.1

# ── Graph visualization ──────────────────────────────────
networkx>=3.1
matplotlib>=3.7.0

# ── Numeric / ML (optional — for Isolation Forest anomaly detection) ──
numpy>=1.24.0
scikit-learn>=1.3.0

# ── PDF report export (optional) ────────────────────────
reportlab>=4.0.0
